* consts.py (twh): Convert to float for good measure.

+++ consts.py
@@ -8,7 +8,7 @@
 """Constants."""
 
 # Conversion factor between MWh and TWh.
-twh = pow(10, 6)
+twh = float(pow(10, 6))
 
 # Default non-synchronous penetration limit.
 nsp_limit = 0.75
